Move allowed props to separate file
