+++ src/handlers/gestureHandlers.ts
@@ -8,6 +8,16 @@ import createHandler from './createHandler';
 import PlatformConstants from '../PlatformConstants';
 import { State } from '../State';
 import { ValueOf } from '../typeUtils';
+import {
+  baseGestureHandlerProps,
+  tapGestureHandlerProps,
+  panGestureHandlerProps,
+  panGestureHandlerCustomNativeProps,
+  longPressGestureHandlerProps,
+  flingGestureHandlerProps,
+  forceTouchGestureHandlerProps,
+  managePanProps,
+} from './allowedProps';
 
 export interface GestureEventPayload {
   handlerTag: number;
@@ -71,23 +81,6 @@ export type BaseGestureHandlerProps<
   ) => void;
 };
 
-export const baseProps = [
-  'id',
-  'enabled',
-  'minPointers',
-  'waitFor',
-  'simultaneousHandlers',
-  'shouldCancelWhenOutside',
-  'hitSlop',
-  'onBegan',
-  'onFailed',
-  'onCancelled',
-  'onActivated',
-  'onEnded',
-  'onGestureEvent',
-  'onHandlerStateChange',
-] as const;
-
 export type TapGestureHandlerEventPayload = {
   x: number;
   y: number;
@@ -114,14 +107,8 @@ export const TapGestureHandler = createHandler<
 >({
   name: 'TapGestureHandler',
   allowedProps: [
-    ...baseProps,
-    'maxDurationMs',
-    'maxDelayMs',
-    'numberOfTaps',
-    'maxDeltaX',
-    'maxDeltaY',
-    'maxDist',
-    'minPointers',
+    ...baseGestureHandlerProps,
+    ...tapGestureHandlerProps,
   ] as const,
   config: {},
 });
@@ -146,7 +133,10 @@ export const FlingGestureHandler = createHandler<
   FlingGestureHandlerEventPayload
 >({
   name: 'FlingGestureHandler',
-  allowedProps: [...baseProps, 'numberOfPointers', 'direction'] as const,
+  allowedProps: [
+    ...baseGestureHandlerProps,
+    ...flingGestureHandlerProps,
+  ] as const,
   config: {},
 });
 
@@ -188,10 +178,8 @@ export const ForceTouchGestureHandler = PlatformConstants?.forceTouchAvailable
     >({
       name: 'ForceTouchGestureHandler',
       allowedProps: [
-        ...baseProps,
-        'minForce',
-        'maxForce',
-        'feedbackOnActivation',
+        ...baseGestureHandlerProps,
+        ...forceTouchGestureHandlerProps,
       ] as const,
       config: {},
     })
@@ -221,178 +209,12 @@ export const LongPressGestureHandler = createHandler<
   LongPressGestureHandlerEventPayload
 >({
   name: 'LongPressGestureHandler',
-  allowedProps: [...baseProps, 'minDurationMs', 'maxDist'] as const,
+  allowedProps: [
+    ...baseGestureHandlerProps,
+    ...longPressGestureHandlerProps,
+  ] as const,
   config: {},
 });
-
-function validatePanGestureHandlerProps(props: PanGestureHandlerProps) {
-  if (props.minDeltaX && props.activeOffsetX) {
-    throw new Error(
-      `It's not supported use minDeltaX with activeOffsetXStart or activeOffsetXEnd`
-    );
-  }
-  if (props.maxDeltaX && props.failOffsetX) {
-    throw new Error(
-      `It's not supported use minDeltaX with activeOffsetXStart or activeOffsetXEnd`
-    );
-  }
-  if (props.minDeltaY && props.activeOffsetY) {
-    throw new Error(
-      `It's not supported use minDeltaX with activeOffsetYStart or activeOffsetYEnd`
-    );
-  }
-  if (props.maxDeltaY && props.failOffsetY) {
-    throw new Error(
-      `It's not supported use minDeltaX with activeOffsetYStart or activeOffsetYEnd`
-    );
-  }
-  if (
-    Array.isArray(props.activeOffsetX) &&
-    (props.activeOffsetX[0] > 0 || props.activeOffsetX[1] < 0)
-  ) {
-    throw new Error(
-      `First element of activeOffsetX should be negative, a the second one should be positive`
-    );
-  }
-
-  if (
-    Array.isArray(props.activeOffsetY) &&
-    (props.activeOffsetY[0] > 0 || props.activeOffsetY[1] < 0)
-  ) {
-    throw new Error(
-      `First element of activeOffsetY should be negative, a the second one should be positive`
-    );
-  }
-
-  if (
-    Array.isArray(props.failOffsetX) &&
-    (props.failOffsetX[0] > 0 || props.failOffsetX[1] < 0)
-  ) {
-    throw new Error(
-      `First element of failOffsetX should be negative, a the second one should be positive`
-    );
-  }
-
-  if (
-    Array.isArray(props.failOffsetY) &&
-    (props.failOffsetY[0] > 0 || props.failOffsetY[1] < 0)
-  ) {
-    throw new Error(
-      `First element of failOffsetY should be negative, a the second one should be positive`
-    );
-  }
-}
-
-function transformPanGestureHandlerProps(props: PanGestureHandlerProps) {
-  type InternalPanGHKeys =
-    | 'activeOffsetXStart'
-    | 'activeOffsetXEnd'
-    | 'failOffsetXStart'
-    | 'failOffsetXEnd'
-    | 'activeOffsetYStart'
-    | 'activeOffsetYEnd'
-    | 'failOffsetYStart'
-    | 'failOffsetYEnd';
-  type PanGestureHandlerInternalProps = PanGestureHandlerProps &
-    Partial<Record<InternalPanGHKeys, number>>;
-
-  const res: PanGestureHandlerInternalProps = { ...props };
-  if (props.minDeltaX !== undefined) {
-    delete res.minDeltaX;
-    res.activeOffsetXStart = -props.minDeltaX;
-    res.activeOffsetXEnd = props.minDeltaX;
-  }
-  if (props.maxDeltaX !== undefined) {
-    delete res.maxDeltaX;
-    res.failOffsetXStart = -props.maxDeltaX;
-    res.failOffsetXEnd = props.maxDeltaX;
-  }
-  if (props.minOffsetX !== undefined) {
-    delete res.minOffsetX;
-    if (props.minOffsetX < 0) {
-      res.activeOffsetXStart = props.minOffsetX;
-    } else {
-      res.activeOffsetXEnd = props.minOffsetX;
-    }
-  }
-
-  if (props.minDeltaY !== undefined) {
-    delete res.minDeltaY;
-    res.activeOffsetYStart = -props.minDeltaY;
-    res.activeOffsetYEnd = props.minDeltaY;
-  }
-  if (props.maxDeltaY !== undefined) {
-    delete res.maxDeltaY;
-    res.failOffsetYStart = -props.maxDeltaY;
-    res.failOffsetYEnd = props.maxDeltaY;
-  }
-
-  if (props.minOffsetY !== undefined) {
-    delete res.minOffsetY;
-    if (props.minOffsetY < 0) {
-      res.activeOffsetYStart = props.minOffsetY;
-    } else {
-      res.activeOffsetYEnd = props.minOffsetY;
-    }
-  }
-
-  if (props.activeOffsetX !== undefined) {
-    delete res.activeOffsetX;
-    if (Array.isArray(props.activeOffsetX)) {
-      res.activeOffsetXStart = props.activeOffsetX[0];
-      res.activeOffsetXEnd = props.activeOffsetX[1];
-    } else if (props.activeOffsetX < 0) {
-      res.activeOffsetXStart = props.activeOffsetX;
-    } else {
-      res.activeOffsetXEnd = props.activeOffsetX;
-    }
-  }
-
-  if (props.activeOffsetY !== undefined) {
-    delete res.activeOffsetY;
-    if (Array.isArray(props.activeOffsetY)) {
-      res.activeOffsetYStart = props.activeOffsetY[0];
-      res.activeOffsetYEnd = props.activeOffsetY[1];
-    } else if (props.activeOffsetY < 0) {
-      res.activeOffsetYStart = props.activeOffsetY;
-    } else {
-      res.activeOffsetYEnd = props.activeOffsetY;
-    }
-  }
-
-  if (props.failOffsetX !== undefined) {
-    delete res.failOffsetX;
-    if (Array.isArray(props.failOffsetX)) {
-      res.failOffsetXStart = props.failOffsetX[0];
-      res.failOffsetXEnd = props.failOffsetX[1];
-    } else if (props.failOffsetX < 0) {
-      res.failOffsetXStart = props.failOffsetX;
-    } else {
-      res.failOffsetXEnd = props.failOffsetX;
-    }
-  }
-
-  if (props.failOffsetY !== undefined) {
-    delete res.failOffsetY;
-    if (Array.isArray(props.failOffsetY)) {
-      res.failOffsetYStart = props.failOffsetY[0];
-      res.failOffsetYEnd = props.failOffsetY[1];
-    } else if (props.failOffsetY < 0) {
-      res.failOffsetYStart = props.failOffsetY;
-    } else {
-      res.failOffsetYEnd = props.failOffsetY;
-    }
-  }
-
-  return res;
-}
-
-function managePanProps(props: PanGestureHandlerProps) {
-  if (__DEV__) {
-    validatePanGestureHandlerProps(props);
-  }
-  return transformPanGestureHandlerProps(props);
-}
 
 export type PanGestureHandlerEventPayload = {
   x: number;
@@ -441,32 +263,12 @@ export const PanGestureHandler = createHandler<
 >({
   name: 'PanGestureHandler',
   allowedProps: [
-    ...baseProps,
-    'activeOffsetY',
-    'activeOffsetX',
-    'failOffsetY',
-    'failOffsetX',
-    'minDist',
-    'minVelocity',
-    'minVelocityX',
-    'minVelocityY',
-    'minPointers',
-    'maxPointers',
-    'avgTouches',
-    'enableTrackpadTwoFingerGesture',
+    ...baseGestureHandlerProps,
+    ...panGestureHandlerProps,
   ] as const,
   config: {},
   transformProps: managePanProps,
-  customNativeProps: [
-    'activeOffsetYStart',
-    'activeOffsetYEnd',
-    'activeOffsetXStart',
-    'activeOffsetXEnd',
-    'failOffsetYStart',
-    'failOffsetYEnd',
-    'failOffsetXStart',
-    'failOffsetXEnd',
-  ],
+  customNativeProps: panGestureHandlerCustomNativeProps,
 });
 
 export type PinchGestureHandlerEventPayload = {
@@ -486,7 +288,7 @@ export const PinchGestureHandler = createHandler<
   PinchGestureHandlerEventPayload
 >({
   name: 'PinchGestureHandler',
-  allowedProps: baseProps,
+  allowedProps: baseGestureHandlerProps,
   config: {},
 });
 
@@ -507,6 +309,6 @@ export const RotationGestureHandler = createHandler<
   RotationGestureHandlerEventPayload
 >({
   name: 'RotationGestureHandler',
-  allowedProps: baseProps,
+  allowedProps: baseGestureHandlerProps,
   config: {},
 });
